Summarizer/requirements.txt
lexrank==0.0.1b0
numpy==1.13.3
tqdm==4.21.0
nltk==3.2.5
urlextract==0.8.3
newspaper3k==0.2.6
newspaper==0.1.0.7
forked_path==0.2.3
